App Folder/requirements.txt
pandas
numpy
scikit-learn
sqlalchemy
psycopg2-binary
streamlit
matplotlib
joblib
reportlab
pillow
python-pptx